feat: Add broadcast-safra.txt and docs.txt files
This commit adds two new files, broadcast-safra.txt and docs.txt, to the repository. The broadcast-safra.txt file contains a list of URLs and tokens related to the Safra broadcast feature. The docs.txt file contains information about CPFs, CNPJs, and protocols. These files are necessary for the functionality of the application.